requirements.txt
aws-cdk-lib==2.81.0
constructs>=10.0.0,<11.0.0
aws-cdk-lib>=2.0.0
constructs>=10.0.0
PyYAML>=5.4
python-dotenv==0.19.0
boto3==1.24.0
